Minor refactoring of inline edit commands
