+++ src/main/kotlin/com/sourcegraph/cody/edit/FixupUndoableAction.kt
@@ -10,6 +10,7 @@ import com.intellij.openapi.editor.Document
 import com.intellij.openapi.editor.RangeMarker
 import com.intellij.openapi.util.TextRange
 import com.sourcegraph.cody.agent.protocol.TextEdit
+import com.sourcegraph.cody.edit.sessions.FixupSession
 
 abstract class FixupUndoableAction(
     val session: FixupSession,
