+++ src/main/kotlin/com/sourcegraph/cody/edit/widget/LensAction.kt
@@ -2,7 +2,7 @@ package com.sourcegraph.cody.edit.widget
 
 import com.intellij.openapi.editor.event.EditorMouseEvent
 import com.intellij.ui.JBColor
-import com.sourcegraph.cody.edit.FixupSession
+import com.sourcegraph.cody.edit.sessions.FixupSession
 import java.awt.Font
 import java.awt.FontMetrics
 import java.awt.Graphics2D
